Fixed Webhook for Repository URL that use Env Vars
If a Repository URL is specified with an environment variable, for example, https://somewhere.org/soemthing/${JOB_BASE_NAME}, the Webhook fails to trigger a build because the environment variable is not substituted and results in the URI 'match' failing. Using GitSCM.getParamExpandedRepo to expand the environment variables solves the problem.
A further improvement would be to add getParamExpandedRepos(EnvVars env) to GitSCM to convert all the repos in a Collection at the same time.